Revised testing to all be in one file.
