+++ makefile
@@ -9,16 +9,11 @@ hmap: src/hmap.h src/hmap_test.c src/test_helpers.h
 hash_test: src/hash_test.c src/ahash.h
 	$(CC) $(CFLAGS) src/hash_test.c -o $(OUTDIR)/hash_test
 
-dynarr: src/test_main.c src/test_helpers.h src/dynarr.h src/alloc_fail_tests.c
-	$(CC) $(CFLAGS) src/test_main.c -o $(OUTDIR)/test
-	$(CC) $(CFLAGS) src/alloc_fail_tests.c -o $(OUTDIR)/alloc_fail_tests
+dynarr: src/dynarr_test.c src/test_helpers.h src/dynarr.h 
+	$(CC) $(CFLAGS) src/dynarr_test.c -o $(OUTDIR)/dynarr_test
 
 dynarr_test: dynarr
-	$(OUTDIR)/test
-	echo "Starting allocation failure tests!"
-	echo "---------------------------------------------------------------------"
-	$(OUTDIR)/alloc_fail_tests
-
+	$(OUTDIR)/dynarr_test
 
 outdir:
 	mkdir -p $(OUTDIR)
@@ -30,10 +25,7 @@ hmap_test: hmap
 	$(OUTDIR)/hmap_test
 
 test: tests
-	$(OUTDIR)/test
-	echo "Starting allocation failure tests!"
-	echo "---------------------------------------------------------------------"
-	$(OUTDIR)/alloc_fail_tests
+	$(OUTDIR)/dynarr_test
 	echo "Starting hash_test"
 	echo "---------------------------------------------------------------------"
 	$(OUTDIR)hash_test
